requirements.txt
Flask==1.1.1
h5py==2.10.0
Keras==2.3.1
Pillow==7.0.0
requests==2.23.0
tensorflow==2.1.0

